move cached players to proxy
